+++ lib/pmdtester/builders/pmd_report_builder.rb
@@ -60,22 +60,12 @@ module PmdTester
         # In CI, that's not a problem, because the workspace is always fresh.
         logger.warn "#{@pmd_branch_name}: Reusing already existing #{pmd_dist_target}"
       else
-        logger.info "#{@pmd_branch_name}: Building PMD #{@pmd_version}..."
-        package_cmd = './mvnw clean package' \
-                      ' -Dmaven.test.skip=true' \
-                      ' -Dmaven.javadoc.skip=true' \
-                      ' -Dmaven.source.skip=true' \
-                      ' -Dcheckstyle.skip=true' \
-                      ' -Dpmd.skip=true' \
-                      ' -T1C -B'
-        Cmd.execute_successfully(package_cmd)
-
+        build_pmd_with_maven
         pmd_dist_target, binary_exists = find_pmd_dist_target
         unless binary_exists
           logger.error "#{@pmd_branch_name}: Dist zip not found at #{pmd_dist_target}!"
           raise "No Dist zip found at #{pmd_dist_target}"
         end
-
       end
 
       logger.info "#{@pmd_branch_name}: Extracting the zip"
@@ -101,7 +91,7 @@ module PmdTester
 
     def generate_pmd_report(project)
       error_recovery_options = @error_recovery ? 'PMD_JAVA_OPTS="-Dpmd.error_recovery -ea" ' : ''
-      fail_on_violation = should_use_long_cli_options? ? '--fail-on-violation false' : '-failOnViolation false'
+      fail_on_violation = create_failonviolation_option
       auxclasspath_option = create_auxclasspath_option(project)
       pmd_cmd = "#{error_recovery_options}" \
                 "#{determine_run_path} -d #{project.local_source_path} -f xml " \
@@ -218,6 +208,16 @@ module PmdTester
       auxclasspath_option
     end
 
+    def create_failonviolation_option
+      if pmd7?
+        '--no-fail-on-violation'
+      elsif should_use_long_cli_options?
+        '--fail-on-violation false'
+      else
+        '-failOnViolation false'
+      end
+    end
+
     def pmd7?
       Semver.compare(@pmd_version, '7.0.0-SNAPSHOT') >= 0
     end
@@ -243,6 +243,19 @@ module PmdTester
         logger.debug "#{@pmd_branch_name}: Does the file #{pmd_dist_target} exist? #{binary_exists} (cwd: #{Dir.getwd})"
       end
       [pmd_dist_target, binary_exists]
+    end
+
+    def build_pmd_with_maven
+      logger.info "#{@pmd_branch_name}: Building PMD #{@pmd_version}..."
+      package_cmd = './mvnw clean package' \
+                    ' -Dmaven.test.skip=true' \
+                    ' -Dmaven.javadoc.skip=true' \
+                    ' -Dmaven.source.skip=true' \
+                    ' -Dcheckstyle.skip=true' \
+                    ' -Dpmd.skip=true' \
+                    ' -T1C -B'
+      logger.debug "#{@pmd_branch_name}: maven command: #{package_cmd}"
+      Cmd.execute_successfully(package_cmd)
     end
   end
 end