+++ History.md
@@ -4,6 +4,7 @@
 
 ## Enhancements
 * [#118](https://github.com/pmd/pmd-regression-tester/pull/118): Update js libraries
+* [#119](https://github.com/pmd/pmd-regression-tester/pull/119): Update manual integration test for PMD 7
 * [#120](https://github.com/pmd/pmd-regression-tester/pull/120): Support new PMD 7 binary dist filename
 
 ## Fixed Issues